Inline corporate email template
